+++ crates/miden-objects/src/conversion/mod.rs
@@ -1,10 +1,12 @@
 mod account;
+mod account_file;
 mod account_patch;
 mod asset;
 mod batch;
 mod block;
 mod merkle;
 mod note;
+mod note_file;
 mod primitives;
 mod protocol_config;
 mod transaction;