+++ crates/miden-objects/src/conversion/note.rs
@@ -130,6 +130,12 @@ impl From<&NoteDetails> for proto::note::NoteDetails {
 // NOTE
 // ================================================================================================
 
+impl From<&Note> for proto::note::Note {
+    fn from(note: &Note) -> Self {
+        Self::from(note.clone())
+    }
+}
+
 impl From<Note> for proto::note::Note {
     fn from(note: Note) -> Self {
         let (assets, metadata, recipient, attachments) = note.into_parts();